+++ scm-core/src/main/java/sonia/scm/repository/EventDrivenRepositoryArchiveCheck.java
@@ -25,8 +25,6 @@
 package sonia.scm.repository;
 
 import com.github.legman.Subscribe;
-import sonia.scm.EagerSingleton;
-import sonia.scm.plugin.Extension;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -37,8 +35,6 @@ import java.util.HashSet;
  * {@link RepositoryModificationEvent}s. The initial set of archived repositories is read by
  * {@link EventDrivenRepositoryArchiveCheckInitializer} on startup.
  */
-@Extension
-@EagerSingleton
 public final class EventDrivenRepositoryArchiveCheck implements RepositoryArchivedCheck {
 
   private static final Collection<String> ARCHIVED_REPOSITORIES = Collections.synchronizedSet(new HashSet<>());
